% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VivaxFunctionsRCD.R
\name{calculate_h1_rcd}
\alias{calculate_h1_rcd}
\title{Calculate h1 from aggregated incidence data}
\usage{
calculate_h1_rcd(h, alpha, rho, iota, nu, tau, eta, r, rho2)
}
\arguments{
\item{h}{daily incidence}

\item{alpha}{effective treatment probability}

\item{rho}{observation rate}

\item{iota}{RCD parameter: maximum number of index cases investigated per population per unit of time}

\item{nu}{RCD parameter: number of secondary cases investigated per index case}

\item{tau}{RCD parameter: targeting ratio of the RCD}

\item{eta}{RCD parameter: probability that an investigated cases is detected (symptoms, test sensitivity, etc.)}

\item{r}{blood clearance rate}

\item{rho2}{RCD parameter: observation rate for reactively-detected cases}
}
\value{
A scalar equal to the equilibrium value for h1
}
\description{
Calculate h1 in RCD model for back-calculation when
}
