% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VivaxFunctionsRCD.R
\name{calculate_tau_rcd}
\alias{calculate_tau_rcd}
\title{Calculate the targeting ratio from incidence data with information on RCD}
\usage{
calculate_tau_rcd(h1, h2, alpha, rho, iota, nu, eta, r)
}
\arguments{
\item{h1}{daily incidence, directly detected cases}

\item{h2}{daily incidence, reactively detected cases}

\item{alpha}{effective treatment probability}

\item{rho}{observation rate}

\item{iota}{RCD parameter: maximum number of index cases investigated per population per unit of time}

\item{nu}{RCD parameter: number of secondary cases investigated per index case}

\item{eta}{RCD parameter: probability that an investigated cases is detected (symptoms, test sensitivity, etc.)}

\item{r}{blood clearance rate}
}
\value{
a scalar equal to the targeting ratio tau.
}
\description{
Calculate tau in RCD model when h1 and h2 are available
}
