% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{create_approximation}
\alias{create_approximation}
\title{Creates linear interpolation}
\usage{
create_approximation(x, y, start = 0)
}
\arguments{
\item{x}{time points}

\item{y}{values to interpolate}

\item{start}{rescales the x to x-start (useful to chain interventions properly)}
}
\value{
An interpolation function, like the one created by approxfun (stats package)
}
\description{
Creates linear interpolation
}
