% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VivaxFunctions.R
\name{get_equilibrium_states_vivax}
\alias{get_equilibrium_states_vivax}
\title{Calculate equilibrium state variables}
\usage{
get_equilibrium_states_vivax(
  I,
  lambda,
  r,
  gamma,
  f,
  alpha,
  beta,
  rho,
  delta,
  omega
)
}
\arguments{
\item{I}{proportion of infectious individuals at equilibrium (I0+Il)}

\item{lambda}{transmission rate}

\item{r}{blood clearance rate}

\item{gamma}{liver clearance rate}

\item{f}{relapse frequency}

\item{alpha}{effective treatment probability}

\item{beta}{probability of liver clearance}

\item{rho}{observation rate}

\item{delta}{importation rate}

\item{omega}{vector control intensity. If vector control is not the purpose of the analysis, omega can be fixed to 1.}
}
\value{
A list with the equilibrium states (I0, Il, S0 and Sl)
}
\description{
Calculates equilibrium state variables I0, Il, S0 and Sl based on I and model parameters
}
