% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VivaxFunctionsDelays.R
\name{get_equilibrium_states_vivax_delay}
\alias{get_equilibrium_states_vivax_delay}
\title{Calculate equilibrium state variables, in model with delay to treatment}
\usage{
get_equilibrium_states_vivax_delay(
  I,
  lambda,
  r,
  gamma,
  f,
  alpha,
  beta,
  sigma,
  rho,
  delta,
  omega
)
}
\arguments{
\item{I}{proportion of infectious individuals at equilibrium (U0+Ul)}

\item{lambda}{transmission rate}

\item{r}{blood clearance rate}

\item{gamma}{liver clearance rate}

\item{f}{relapse frequency}

\item{alpha}{effective treatment probability}

\item{beta}{probability of liver clearance}

\item{sigma}{delay to treatment (1/duration of the delay)}

\item{rho}{observation rate}

\item{delta}{importation rate}

\item{omega}{vector control intensity. If vector control is not the purpose of the analysis, omega can be fixed to 1.}
}
\value{
A list with the equilibrium states (U0, Ul, S0 and Sl)
}
\description{
Calculates equilibrium state variables U0, Ul, T0, Tl, S0 and Sl based on I and model parameters
}
