% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VivaxFunctionsRCD.R
\name{get_equilibrium_states_vivax_rcd}
\alias{get_equilibrium_states_vivax_rcd}
\title{Calculate equilibrium state variables}
\usage{
get_equilibrium_states_vivax_rcd(
  I,
  lambda,
  r,
  gamma,
  f,
  alpha,
  beta,
  rho,
  delta,
  omega,
  iota_star,
  nu,
  tau,
  eta,
  kappa
)
}
\arguments{
\item{I}{proportion of infectious individuals at equilibrium (I0+Il)}

\item{lambda}{transmission rate}

\item{r}{blood clearance rate}

\item{gamma}{liver clearance rate}

\item{f}{relapse frequency}

\item{alpha}{effective treatment probability}

\item{beta}{probability of liver clearance}

\item{rho}{observation rate}

\item{delta}{importation rate}

\item{omega}{vector control intensity. If vector control is not the purpose of the analysis, omega can be fixed to 1.}

\item{iota_star}{RCD parameter: maximum number of index cases investigated per population per unit of time at equilibrium}

\item{nu}{RCD parameter: number of secondary cases investigated per index case}

\item{tau}{RCD parameter: targeting ratio of the RCD}

\item{eta}{RCD parameter: probability that an investigated cases is detected (symptoms, test sensitivity, etc.)}

\item{kappa}{common part in rho and alpha}
}
\value{
A list with the equilibrium states (I0, Il, S0 and Sl)
}
\description{
Calculates equilibrium state variables I0, Il, S0 and Sl based on I and model parameters
}
