% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VivaxFunctions.R
\name{get_prop_relapse}
\alias{get_prop_relapse}
\title{Calculate proportion of relapses}
\usage{
get_prop_relapse(lambda, h, r, f, gamma, alpha, beta, rho, p, omega)
}
\arguments{
\item{lambda}{transmission rate}

\item{h}{daily incidence}

\item{r}{blood clearance rate}

\item{f}{relapse frequency}

\item{gamma}{liver clearance rate}

\item{alpha}{effective treatment probability}

\item{beta}{probability of liver clearance}

\item{rho}{observation rate}

\item{p}{proportion of imported cases}

\item{omega}{vector control intensity. If vector control is not the purpose of the analysis, omega can be fixed to 1.}
}
\value{
A scalar giving the proportion of relapses
}
\description{
Calculates the proportion of new infections due to relapse at equilibrium
}
