% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VivaxFunctions.R
\name{get_r0_vivax}
\alias{get_r0_vivax}
\title{R0 calculation}
\usage{
get_r0_vivax(lambda, f, r, gamma)
}
\arguments{
\item{lambda}{transmission parameter}

\item{f}{relapse frequency}

\item{r}{blood clearance rate}

\item{gamma}{liver clearance rate}
}
\value{
A scalar R0 value
}
\description{
Calculation of R0 in vivax model
}
