% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VivaxFunctions.R
\name{get_rc_vivax}
\alias{get_rc_vivax}
\title{Rc calculation, with case management and vector control}
\usage{
get_rc_vivax(lambda, f, r, gamma, alpha, beta, omega = 1)
}
\arguments{
\item{lambda}{transmission parameter}

\item{f}{relapse frequency}

\item{r}{blood clearance rate}

\item{gamma}{liver clearance rate}

\item{alpha}{effective treatment probability}

\item{beta}{probability of liver clearance}

\item{omega}{intensity of vector control. If vector control is not the purpose of the analysis, omega can be fixed to 1.}
}
\value{
A scalar Rc value
}
\description{
Calculation of Rc in vivax model
}
