% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VivaxFunctions.R
\name{get_validity_condition_vivax}
\alias{get_validity_condition_vivax}
\title{Checks validity condition}
\usage{
get_validity_condition_vivax(I, r, f, gamma, alpha, beta, delta)
}
\arguments{
\item{I}{proportion of infectious individuals at equilibrium (I0+Il)}

\item{r}{blood clearance rate}

\item{f}{relapse frequency}

\item{gamma}{liver clearance rate}

\item{alpha}{effective treatment probability}

\item{beta}{probability of liver clearance}

\item{delta}{importation rate}
}
\value{
A scalar giving the condition. If negative, the solution exists. If positive there is no positive solution for lambda.
}
\description{
Check if validity condition for transmission rate calculation
based on equilibrium prevalence is verified.
}
