% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{incidence_day2year}
\alias{incidence_day2year}
\title{Converts daily incidence to annual}
\usage{
incidence_day2year(h)
}
\arguments{
\item{h}{daily incidence (cases per 1 person day)}
}
\value{
Scalar representing annual incidence (cases per 1000 person year)
incidence=h\emph{365}1000
}
\description{
Converts daily incidence to annual
}
