% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VivaxFunctionsSto.R
\name{model_sto_vivax_delay_rcd_no_referral}
\alias{model_sto_vivax_delay_rcd_no_referral}
\title{Reaction associated with the vivax model with delays and RCD, in stochastic version}
\usage{
model_sto_vivax_delay_rcd_no_referral()
}
\value{
A list with 2 objects:
model: the stochastic model compiled (as in the TiPS package)
reactions_incidence: a list of vectors indicating which reactions correspond to the incidence variables hh and hhl
}
\description{
This function creates the stochastic model for vivax dynamics with delays to treatment and reactive case detection (without referral to health facilities).
}
