% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VivaxFunctions.R
\name{ode_vivax_cm_importation_vc}
\alias{ode_vivax_cm_importation_vc}
\title{ODE vivax model with case management, importation and vector control}
\usage{
ode_vivax_cm_importation_vc(t, y, parameters)
}
\arguments{
\item{t}{time}

\item{y}{state variables}

\item{parameters}{model parameters}
}
\value{
a list of the state variables for all time steps
}
\description{
Defines the equations of the model
}
