% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VivaxFunctionsMDA.R
\name{ode_vivax_delay_mda}
\alias{ode_vivax_delay_mda}
\title{ODE vivax model with case management, importation and vector control, including delay to treatment, during MDA prophylaxis}
\usage{
ode_vivax_delay_mda(t, y, parameters)
}
\arguments{
\item{t}{time}

\item{y}{state variables}

\item{parameters}{model parameters}
}
\value{
a list of the state variables for all time steps
}
\description{
Defines the equations of the model
}
