% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VivaxFunctionsRCD.R
\name{ode_vivax_rcd}
\alias{ode_vivax_rcd}
\title{ODE vivax model with case management, RCD, importation and vector control, including delay to treatment}
\usage{
ode_vivax_rcd(t, y, parameters)
}
\arguments{
\item{t}{time}

\item{y}{state variables}

\item{parameters}{model parameters}
}
\value{
a list of the state variables for all time steps
}
\description{
Defines the equations of the model
}
