% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{sample_uncertainty_incidence_import}
\alias{sample_uncertainty_incidence_import}
\title{Samples uncertainty in incidence and importation data}
\usage{
sample_uncertainty_incidence_import(df, ndraw = 100)
}
\arguments{
\item{df}{dataframe containing the variables "cases" (reported cases),
"cases_local" (number of reported cases) and "population" (population size)}

\item{ndraw}{number of samples}
}
\value{
A dataframe with ndraw replicates of the original dataframe df, containing 4 additional variables.\cr
h is the sampled daily incidence (cases per person day), \cr
prop_import is the sampled proportion of imported cases
}
\description{
Samples uncertainty in incidence and importation data
}
