% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanity_checks.R
\name{sanity_checks_inputs_simulate}
\alias{sanity_checks_inputs_simulate}
\title{Check on the database input for simulating the model on a database of values}
\usage{
sanity_checks_inputs_simulate(df, delay, rcd, mda, sto, rcd_at_baseline)
}
\arguments{
\item{df}{a dataframe containing the data, with one column called h containing the daily incidence
and one variable called prop_import containing the proportion of imported cases among new infections.
additional optional variables are alpha (effective care), beta (proportion of liver stage cure),
rho (reporting rate) and omega (intensity of vector control)}

\item{delay}{a boolean indicating if the model including delays in treatment should be used.}

\item{rcd}{a boolean indicating if the model including reactive case detection should be used.}

\item{mda}{a boolean indicating if the model including mass drug administration (MDA) prophylaxis should be used.}

\item{sto}{a boolean indicating if the stochastic version of the model should be used.}

\item{rcd_at_baseline}{a boolean indicating if the model was calibrated using the RCD model (i.e. there is some RCD at baseline already). Default (FALSE) is the model without RCD at baseline}
}
\value{
A dataframe containing all required variables
}
\description{
Checks the presence of required variables and adds default values for missing variables.
}
