% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VivaxFunctionsDelays.R
\name{simulate_vivax_delay_ode}
\alias{simulate_vivax_delay_ode}
\title{Simulate vivax model with delay to treatment}
\usage{
simulate_vivax_delay_ode(
  parameters,
  ODEmodel = ode_vivax_delay,
  maxtime = 1465,
  year = FALSE
)
}
\arguments{
\item{parameters}{model parameters}

\item{ODEmodel}{ODE model to be simulated}

\item{maxtime}{maximal time step}

\item{year}{if TRUE, aggregates outputs per year. if FALSE, returns daily output}
}
\value{
A dataframe containing the simulated model
}
\description{
Simulates a draw of the chosen vivax ODE model
}
