% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VivaxFunctionsSto.R
\name{simulate_vivax_delay_sto}
\alias{simulate_vivax_delay_sto}
\title{Simulate vivax model with delay to treatment, with stochastic model}
\usage{
simulate_vivax_delay_sto(
  parameters,
  STOmodel = model_sto_vivax_delay(),
  maxtime = 1465,
  year = FALSE,
  sto_method = "exact",
  runs = 1,
  seeds = NULL
)
}
\arguments{
\item{parameters}{model parameters}

\item{STOmodel}{stochastic model to be simulated (reactions)}

\item{maxtime}{maximal time step}

\item{year}{if TRUE, aggregates outputs per year. if FALSE, returns daily output}

\item{sto_method}{a scalarindicating which simulation method is used.
Default ("exact") is Gillespie's direct method. Other options are "approximate" (tau-leap) or "mixed".
cf. the documentation of the TiPS package for more information.}

\item{runs}{number of draws of the stochastic model}

\item{seeds}{a vector of the length of runs containing the seeds for each simulation (don't use "0" which has another use in TiPS)}
}
\value{
A dataframe containing the simulated model
}
\description{
Simulates a draw of the chosen vivax stochastic model
}
