% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VivaxFunctionsMDA.R
\name{simulate_vivax_mda_ode}
\alias{simulate_vivax_mda_ode}
\title{Simulate vivax model with MDA}
\usage{
simulate_vivax_mda_ode(
  parameters,
  ODEmodel = ode_vivax_cm_importation_vc,
  ODEmodel_mda = ode_vivax_mda,
  maxtime = 1465,
  year = FALSE
)
}
\arguments{
\item{parameters}{model parameters}

\item{ODEmodel}{ODE model to be simulated}

\item{ODEmodel_mda}{ODE model with MDA to be simulated (during prophylaxis)}

\item{maxtime}{maximal time step}

\item{year}{if TRUE, aggregates outputs per year. if FALSE, returns daily output}
}
\value{
A dataframe containing the simulated model
}
\description{
Simulates a draw of the chosen vivax ODE model
}
