% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VivaxFunctionsDelays.R
\name{solve_lambda_vivax_delay}
\alias{solve_lambda_vivax_delay}
\title{Transmission rate calculation, in model with delay to treatment}
\usage{
solve_lambda_vivax_delay(
  h,
  r,
  f,
  gamma,
  alpha,
  beta,
  rho,
  sigma,
  p,
  omega,
  return.all = FALSE
)
}
\arguments{
\item{h}{daily incidence}

\item{r}{blood clearance rate}

\item{f}{relapse frequency}

\item{gamma}{liver clearance rate}

\item{alpha}{effective treatment probability}

\item{beta}{probability of liver clearance}

\item{rho}{observation rate}

\item{sigma}{delay to treatment (1/duration of the delay)}

\item{p}{proportion of imported cases}

\item{omega}{vector control intensity. If vector control is not the purpose of the analysis, omega can be fixed to 1.}

\item{return.all}{if TRUE, returns lambda, I and delta. If FALSE, returns only lambda}
}
\value{
If return.all=TRUE, returns a list with lambda, I and delta. If return.all=FALSE, returns only a lambda value
}
\description{
Solve the equation for lambda, vivax model with CM, vector control and importation
}
