% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VivaxFunctionsRCD.R
\name{solve_lambda_vivax_rcd}
\alias{solve_lambda_vivax_rcd}
\title{Transmission rate calculation}
\usage{
solve_lambda_vivax_rcd(
  h,
  h1,
  r,
  f,
  gamma,
  alpha,
  beta,
  rho,
  p,
  omega,
  iota,
  nu,
  tau,
  eta,
  return.all = FALSE
)
}
\arguments{
\item{h}{daily incidence}

\item{h1}{daily incidence, directly detected cases}

\item{r}{blood clearance rate}

\item{f}{relapse frequency}

\item{gamma}{liver clearance rate}

\item{alpha}{effective treatment probability}

\item{beta}{probability of liver clearance}

\item{rho}{observation rate}

\item{p}{proportion of imported cases}

\item{omega}{vector control intensity. If vector control is not the purpose of the analysis, omega can be fixed to 1.}

\item{iota}{RCD parameter: maximum number of index cases investigated per population per unit of time}

\item{nu}{RCD parameter: number of secondary cases investigated per index case}

\item{tau}{RCD parameter: targeting ratio of the RCD}

\item{eta}{RCD parameter: probability that an investigated cases is detected (symptoms, test sensitivity, etc.)}

\item{return.all}{if TRUE, returns lambda, I and delta. If FALSE, returns only lambda}
}
\value{
If return.all=TRUE, returns a list with lambda, I and delta. If return.all=FALSE, returns only a lambda value
}
\description{
Solve the equation for lambda, vivax with CM, vector control, RCD and importation
}
