% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{varying_tau}
\alias{varying_tau}
\title{Function for time varying tau}
\usage{
varying_tau(nu, pr, N = 10000)
}
\arguments{
\item{nu}{time}

\item{pr}{prevalence}

\item{N}{population}
}
\value{
A scalar value for tau
}
\description{
From Chitnis et al. 2019 (supplementary information)
}
