% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{vector_control_exponential_decay}
\alias{vector_control_exponential_decay}
\title{Creates vector control exponential decay}
\usage{
vector_control_exponential_decay(
  initial_omega,
  half_life,
  every_x_years = 3,
  maxtime
)
}
\arguments{
\item{initial_omega}{the reduction in vectorial capacity at time=0}

\item{half_life}{half life of the exponential decay (in years)}

\item{every_x_years}{frequency at which the intervention is renewed (default is every 3 years)}

\item{maxtime}{maximum time point to be included in the database (from 0 to maxtime days)}
}
\value{
An interpolation function, like the one created by approxfun (stats package)
}
\description{
Creates vector control exponential decay
}
