% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VivaxModelR-package.R
\docType{package}
\name{VivaxModelR-package}
\alias{VivaxModelR}
\alias{VivaxModelR-package}
\title{VivaxModelR: Simulation of Plasmodium Vivax dynamics}
\description{
This packages contains functions to simulate Plasmodium Vivax dynamics using a compartmental model.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://swisstph.github.io/VivaxModelR/}
}

}
\author{
\strong{Maintainer}: Clara Champagne \email{clara.champagne@swisstph.ch}

Authors:
\itemize{
  \item Maximilian Gerhards
  \item Justin Lana
  \item Arnaud Le Menach
  \item Emilie Pothin \email{emilie.pothin@swisstph.ch}
}

}
\keyword{internal}
