% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateFromData.R
\name{format_data_simulation}
\alias{format_data_simulation}
\title{Formating intervention parameters within a database}
\usage{
format_data_simulation(df, intervention_object)
}
\arguments{
\item{df}{a dataframe containing at least 3 variables : alpha, beta and omega}

\item{intervention_object}{an named list containing the intervention description. It should have the follwing structure:
list(intervention_name="string", "alpha.new"=NA, "beta.new"=NA, "omega.new"=NA )
where NA can be replaced by scalars or kepts as such.}
}
\value{
A dataframe with the required input parameters for future simulation
}
\description{
Helper function to format dataframes
}
\details{
If alpha.new is not provided in intervention_object it is equal to alpha.
If beta.new is not provided in intervention_object it is equal to beta.
If omega.new is not provided in intervention_object it is equal to omega.
}
