% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VivaxModelR-package.R
\docType{package}
\name{VivaxModelR-package}
\alias{VivaxModelR}
\alias{VivaxModelR-package}
\title{VivaxModelR: Simulation of Plasmodium Vivax dynamics}
\description{
This packages contains functions to simulate Plasmodium Vivax dynamics using a compartmental model.
}
\author{
\strong{Maintainer}: Clara Champagne \email{clara.champagne@swisstph.ch}

Authors:
\itemize{
  \item Emilie Pothin \email{emilie.pothin@swisstph.ch}
  \item Maximilian Gerhards \email{maximilian.gerhards@gmx.de}
}

}
\keyword{internal}
